Crate mailpot_web

Re-exports

pub use cal::calendarize;

pub use typed_paths::tsr::RouterExt;

pub use auth::*;

pub use cal::*;

pub use help::*;

pub use lists::*;

pub use minijinja_utils::*;

pub use settings::*;

pub use typed_paths::*;

pub use utils::*;

Modules

anyhow

github crates-io docs-rs

auth

auth_impls 🔒

cal

help

lists

log

A lightweight logging facade.

mail

Types for processing new posts: PostFilter, ListContext, MailJob and PostAction.

melib

A crate that performs mail client operations such as

minijinja_utils

Utils for templates with the [minijinja] crate.

models

Database models: MailingList, ListOwner, ListSubscription, PostPolicy, SubscriptionPolicy and Post.

postfix

Generate configuration for the postfix mail server.

rusqlite

Rusqlite is an ergonomic wrapper for using SQLite from Rust.

serde

Serde

serde_json

Serde JSON

settings

submission

Submit e-mail through SMTP.

typed_paths

utils

Structs

AppState

AuthLayer

Layer that provides session-based authentication via [AuthContext].

AuthMemoryStore

An ephemeral store, useful for testing and demonstration purposes.

Configuration

The configuration for the mailpot database and the mail server.

Connection

A connection to a mailpot database.

DbVal

A database entry and its primary key. Derefs to its inner type.

Extension

Extractor and response for extensions.

Form

URL encoded extractor and response.

Html

An HTML response.

MailtoAddress

A mailto: value.

MemoryStore

in-memory session store

Path

Extractor that will get captures from the URL and parse them using serde.

Query

Extractor that deserializes query strings into some type.

QueueEntry

A queue entry.

ReadableSession

An extractor which provides a readable session. Sessions may have many readers.

Redirect

Response that redirects the request to another location.

Request

Represents an HTTP request.

RequireAuthorizationLayer

A wrapper around [tower_http::auth::RequireAuthorizationLayer] which provides login authorization.

Response

Represents an HTTP response

ResponseError

Router

The router type for composing handlers and services.

SessionLayer

Layer that provides cookie-based sessions.

State

Extractor for state.

StatusCode

An HTTP status code (status-code in RFC 7230 et al.).

Template

A named template.

TemplateRenderContext

Helper type for Connection::send_reply_with_list_template.

WritableSession

An extractor which provides a writable session. Sessions may have only one writer.

Enums

ErrorKind

The kind of an error.

Queue

In-database queues of mail.

Result

Result is a type that represents either success (Ok) or failure (Err).

SendMail

How to send e-mail.

Traits

AuthUser

A trait which defines methods that allow an arbitrary user type to be authenticated.

ChainedError

This trait is implemented on all the errors generated by the error_chain macro.

Context

Provides the context method for Result.

Handler

Trait for async functions that can be used to handle requests.

IntoResponse

Trait for generating responses.

IntoResponseError

IntoResponseErrorResult

OptionalExtension

See the method documentation.

ResultExt

Additional methods for Result, for easy interaction with this crate.

TypedPath

A type safe path.

Functions

get

Route GET requests to the given handler.

post

Route POST requests to the given handler.

Type Definitions

AuthContext

RequireAuth

SecretVec

Vec types containing secret value

Derive Macros

TypedPath

Derive an implementation of axum_extra::routing::TypedPath.