Module mailpot_web::mail::message_filters

Filters to pass each mailing list post through. Filters are functions that implement the PostFilter trait that can:

transform post content.

modify the final PostAction to take.

modify the final scheduled jobs to perform. (See MailJob).

Filters are executed in sequence like this:

let result = filters
    .into_iter()
    .fold(Ok((&mut post, &mut list_ctx)), |p, f| {
        p.and_then(|(p, c)| f.feed(p, c))
    });

so the processing stops at the first returned error.

Structs

AddListHeaders

Add List-* headers

ArchivedAtLink

Adds Archived-At field, if configured.

FinalizeRecipients

Assuming there are no more changes to be done on the post, it finalizes which list subscriptions will receive the post in post.action field.

FixCRLF

Ensure message contains only \r\n line terminators, required by SMTP.

PostRightsCheck

Check that submitter can post to list, for now it accepts everything.

Traits

PostFilter

Filter that modifies and/or verifies a post candidate. On rejection, return a string describing the error and optionally set post.action to Reject or Defer