Struct mailpot_web::minijinja_utils::MailingList

pub struct MailingList {
    pub pk: i64,
    pub name: String,
    pub id: String,
    pub address: String,
    pub description: Option<String>,
    pub topics: Vec<String>,
    pub archive_url: Option<String>,
    pub inner: DbVal<MailingList>,
    pub is_description_html_safe: bool,
}

Fields
pk: i64name: Stringid: Stringaddress: Stringdescription: Option<String>topics: Vec<String>archive_url: Option<String>inner: DbVal<MailingList>is_description_html_safe: bool
Implementations

impl MailingList

pub fn set_safety<O: Borrow<ListOwner>>( &mut self, owners: &[O], administrators: &[String] )

Set whether it’s safe to not escape the list’s description field.

If anyone can display arbitrary html in the server, that’s bad.

Note: uses Borrow so that it can use both DbVal<ListOwner> and ListOwner slices.

Examples found in repository?

web/src/main.rs (line 204)

184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226

async fn root(
    mut session: WritableSession,
    auth: AuthContext,
    State(state): State<Arc<AppState>>,
) -> Result<Html<String>, ResponseError> {
    let db = Connection::open_db(state.conf.clone())?;
    let lists_values = db.lists()?;
    let lists = lists_values
        .iter()
        .map(|list| {
            let months = db.months(list.pk)?;
            let posts = db.list_posts(list.pk, None)?;
            let newest = posts.last().and_then(|p| {
                chrono::Utc
                    .timestamp_opt(p.timestamp as i64, 0)
                    .earliest()
                    .map(|d| d.to_string())
            });
            let list_owners = db.list_owners(list.pk)?;
            let mut list_obj = MailingList::from(list.clone());
            list_obj.set_safety(list_owners.as_slice(), &state.conf.administrators);
            Ok(minijinja::context! {
                newest,
                posts => &posts,
                months => &months,
                list => Value::from_object(list_obj),
            })
        })
        .collect::<Result<Vec<_>, mailpot::Error>>()?;
    let crumbs = vec![Crumb {
        label: "Home".into(),
        url: "/".into(),
    }];

    let context = minijinja::context! {
        page_title => Option::<&'static str>::None,
        lists => &lists,
        current_user => auth.current_user,
        messages => session.drain_messages(),
        crumbs => crumbs,
    };
    Ok(Html(TEMPLATES.get_template("lists.html")?.render(context)?))
}