Struct mailpot_web::Connection

pub struct Connection {
    pub connection: Connection,
    conf: Configuration,
}

A connection to a mailpot database.

Fields
connection: Connection

The rusqlite connection handle.

conf: Configuration
Implementations

impl Connection

pub fn fetch_templates(&self) -> Result<Vec<DbVal<Template>, Global>, Error>

Fetch all.

pub fn fetch_template( &self, template: &str, list_pk: Option<i64> ) -> Result<Option<DbVal<Template>>, Error>

Fetch a named template.

pub fn add_template(&self, template: Template) -> Result<DbVal<Template>, Error>

Insert a named template.

pub fn remove_template( &self, template: &str, list_pk: Option<i64> ) -> Result<Template, Error>

Remove a named template.

impl Connection

pub fn insert_to_queue( &self, entry: QueueEntry ) -> Result<DbVal<QueueEntry>, Error>

Insert a received email into a queue.

pub fn queue( &self, queue: Queue ) -> Result<Vec<DbVal<QueueEntry>, Global>, Error>

Fetch all queue entries.

pub fn delete_from_queue( &mut self, queue: Queue, index: Vec<i64, Global> ) -> Result<Vec<QueueEntry, Global>, Error>

Delete queue entries returning the deleted values.

impl Connection

pub fn insert_to_error_queue( &self, list_pk: Option<i64>, env: &Envelope, raw: &[u8], reason: String ) -> Result<i64, Error>

Insert a received email into the error queue.

pub fn error_queue(&self) -> Result<Vec<DbVal<Value>, Global>, Error>

Fetch all error queue entries.

pub fn delete_from_error_queue( &mut self, index: Vec<i64, Global> ) -> Result<(), Error>

Delete error queue entries.

impl Connection

pub fn insert_post( &self, list_pk: i64, message: &[u8], env: &Envelope ) -> Result<i64, Error>

Insert a mailing list post into the database.

pub fn post( &mut self, env: &Envelope, raw: &[u8], _dry_run: bool ) -> Result<(), Error>

Process a new mailing list post.

pub fn request( &mut self, list: &DbVal<MailingList>, request: ListRequest, env: &Envelope, raw: &[u8] ) -> Result<(), Error>

Process a new mailing list request.

pub fn months(&self, list_pk: i64) -> Result<Vec<String, Global>, Error>

Fetch all year and month values for which at least one post exists in yyyy-mm format.

pub fn list_post_by_message_id( &self, list_pk: i64, message_id: &str ) -> Result<Option<DbVal<Post>>, Error>

Find a post by its Message-ID email header.

pub fn send_reply_with_list_template<'ctx, F>( &self, render_context: TemplateRenderContext<'ctx, F>, recipients: impl Iterator<Item = Cow<'ctx, Address>> ) -> Result<(), Error>where F: Fn() -> Template,

Helper function to send a template reply.

impl Connection

pub fn list_subscriptions( &self, pk: i64 ) -> Result<Vec<DbVal<ListSubscription>, Global>, Error>

Fetch all subscriptions of a mailing list.

pub fn list_subscription( &self, list_pk: i64, pk: i64 ) -> Result<DbVal<ListSubscription>, Error>

Fetch mailing list subscription.

pub fn list_subscription_by_address( &self, list_pk: i64, address: &str ) -> Result<DbVal<ListSubscription>, Error>

Fetch mailing list subscription by their address.

pub fn add_subscription( &self, list_pk: i64, new_val: ListSubscription ) -> Result<DbVal<ListSubscription>, Error>

Add subscription to mailing list.

pub fn add_candidate_subscription( &mut self, list_pk: i64, new_val: ListSubscription ) -> Result<DbVal<ListCandidateSubscription>, Error>

Create subscription candidate.

pub fn accept_candidate_subscription( &mut self, pk: i64 ) -> Result<DbVal<ListSubscription>, Error>

Accept subscription candidate.

pub fn remove_subscription( &self, list_pk: i64, address: &str ) -> Result<(), Error>

Remove a subscription by their address.

pub fn update_subscription( &mut self, change_set: ListSubscriptionChangeset ) -> Result<(), Error>

Update a mailing list subscription.

pub fn account(&self, pk: i64) -> Result<Option<DbVal<Account>>, Error>

Fetch account by pk.

pub fn account_by_address( &self, address: &str ) -> Result<Option<DbVal<Account>>, Error>

Fetch account by address.

pub fn account_subscriptions( &self, pk: i64 ) -> Result<Vec<DbVal<ListSubscription>, Global>, Error>

Fetch all subscriptions of an account by primary key.

pub fn accounts(&self) -> Result<Vec<DbVal<Account>, Global>, Error>

Fetch all accounts.

pub fn add_account(&self, new_val: Account) -> Result<DbVal<Account>, Error>

Add account.

pub fn remove_account(&self, address: &str) -> Result<(), Error>

Remove an account by their address.

pub fn update_account( &mut self, change_set: AccountChangeset ) -> Result<(), Error>

Update an account.

impl Connection

pub fn list_post_policy( &self, pk: i64 ) -> Result<Option<DbVal<PostPolicy>>, Error>

Fetch the post policy of a mailing list.

pub fn remove_list_post_policy( &self, list_pk: i64, policy_pk: i64 ) -> Result<(), Error>

Remove an existing list policy.

let db = Connection::open_or_create_db(config).unwrap().trusted();
let list_pk = db
    .create_list(MailingList {
        pk: 0,
        name: "foobar chat".into(),
        id: "foo-chat".into(),
        address: "foo-chat@example.com".into(),
        description: None,
        archive_url: None,
    })
    .unwrap()
    .pk;
db.set_list_post_policy(PostPolicy {
    pk: 0,
    list: list_pk,
    announce_only: false,
    subscription_only: true,
    approval_needed: false,
    open: false,
    custom: false,
})
.unwrap();
db.remove_list_post_policy(1, 1).unwrap();

pub fn set_list_post_policy( &self, policy: PostPolicy ) -> Result<DbVal<PostPolicy>, Error>

Set the unique post policy for a list.

impl Connection

pub fn list_subscription_policy( &self, pk: i64 ) -> Result<Option<DbVal<SubscriptionPolicy>>, Error>

Fetch the subscription policy of a mailing list.

pub fn remove_list_subscription_policy( &self, list_pk: i64, policy_pk: i64 ) -> Result<(), Error>

Remove an existing list policy.

let db = Connection::open_or_create_db(config).unwrap().trusted();
let list_pk = db
    .create_list(MailingList {
        pk: 0,
        name: "foobar chat".into(),
        id: "foo-chat".into(),
        address: "foo-chat@example.com".into(),
        description: None,
        archive_url: None,
    })
    .unwrap()
    .pk;
db.set_list_post_policy(PostPolicy {
    pk: 0,
    list: list_pk,
    announce_only: false,
    subscription_only: true,
    approval_needed: false,
    open: false,
    custom: false,
})
.unwrap();
db.remove_list_post_policy(1, 1).unwrap();

pub fn set_list_subscription_policy( &self, policy: SubscriptionPolicy ) -> Result<DbVal<SubscriptionPolicy>, Error>

Set the unique post policy for a list.

impl Connection

pub fn open_db(conf: Configuration) -> Result<Connection, Error>

Creates a new database connection.

Connection supports a limited subset of operations by default (see Connection::untrusted). Use Connection::trusted to remove these limits.

pub fn trusted(self) -> Connection

Removes operational limits from this connection. (see Connection::untrusted)

pub fn untrusted(self) -> Connection

Sets operational limits for this connection.

Allow INSERT, DELETE only for “queue”, “candidate_subscription”, “subscription”.

Allow UPDATE only for “subscription” user facing settings.

Allow INSERT only for “post”.

Allow read access to all tables.

Allow SELECT, TRANSACTION, SAVEPOINT, and the strftime function.

Deny everything else.

pub fn open_or_create_db(conf: Configuration) -> Result<Connection, Error>

Create a database if it doesn’t exist and then open it.

pub fn conf(&self) -> &Configuration

Returns a connection’s configuration.

pub fn load_archives(&self) -> Result<(), Error>

Loads archive databases from Configuration::data_path, if any.

pub fn lists(&self) -> Result<Vec<DbVal<MailingList>, Global>, Error>

Returns a vector of existing mailing lists.

pub fn list(&self, pk: i64) -> Result<Option<DbVal<MailingList>>, Error>

Fetch a mailing list by primary key.

pub fn list_by_id<S>(&self, id: S) -> Result<Option<DbVal<MailingList>>, Error>where S: AsRef<str>,

Fetch a mailing list by id.

pub fn create_list( &self, new_val: MailingList ) -> Result<DbVal<MailingList>, Error>

Create a new list.

pub fn list_posts( &self, list_pk: i64, _date_range: Option<(String, String)> ) -> Result<Vec<DbVal<Post>, Global>, Error>

Fetch all posts of a mailing list.

pub fn list_owners( &self, pk: i64 ) -> Result<Vec<DbVal<ListOwner>, Global>, Error>

Fetch the owners of a mailing list.

pub fn remove_list_owner(&self, list_pk: i64, owner_pk: i64) -> Result<(), Error>

Remove an owner of a mailing list.

pub fn add_list_owner( &self, list_owner: ListOwner ) -> Result<DbVal<ListOwner>, Error>

Add an owner of a mailing list.

pub fn update_list( &mut self, change_set: MailingListChangeset ) -> Result<(), Error>

Update a mailing list.

pub fn list_filters( &self, _list: &DbVal<MailingList> ) -> Vec<Box<dyn PostFilter + 'static, Global>, Global>

Return the post filters of a mailing list.

impl Connection

pub fn new_smtp_connection( &self ) -> Result<Pin<Box<dyn Future<Output = Result<SmtpConnection, Error>> + Send + 'static, Global>>, Error>

Return an SMTP connection handle if the database connection has one configured.

pub async fn submit( smtp_connection: &mut SmtpConnection, message: &QueueEntry, dry_run: bool ) -> impl Future<Output = Result<(), Error>>

Submit queue items from values to their recipients.