Struct mailpot_web::Extension

pub struct Extension<T>(pub T);

Extractor and response for extensions.

As extractor

This is commonly used to share state across handlers.

use axum::{
    Router,
    Extension,
    routing::get,
};
use std::sync::Arc;

// Some shared state used throughout our application
struct State {
    // ...
}

async fn handler(state: Extension<Arc<State>>) {
    // ...
}

let state = Arc::new(State { /* ... */ });

let app = Router::new().route("/", get(handler))
    // Add middleware that inserts the state into all incoming request's
    // extensions.
    .layer(Extension(state));

If the extension is missing it will reject the request with a 500 Internal Server Error response.

As response

Response extensions can be used to share state with middleware.

use axum::{
    Extension,
    response::IntoResponse,
};

async fn handler() -> (Extension<Foo>, &'static str) {
    (
        Extension(Foo("foo")),
        "Hello, World!"
    )
}

#[derive(Clone)]
struct Foo(&'static str);

Tuple Fields
0: T