Struct mailpot_web::SessionLayer

pub struct SessionLayer<Store> {
    store: Store,
    cookie_path: String,
    cookie_name: String,
    cookie_domain: Option<String>,
    persistence_policy: PersistencePolicy,
    session_ttl: Option<Duration>,
    same_site_policy: SameSite,
    http_only: bool,
    secure: bool,
    key: Key,
}

Layer that provides cookie-based sessions.

Fields
store: Storecookie_path: Stringcookie_name: Stringcookie_domain: Option<String>persistence_policy: PersistencePolicysession_ttl: Option<Duration>same_site_policy: SameSitehttp_only: boolsecure: boolkey: Key
Implementations

impl<Store> SessionLayer<Store>where Store: SessionStore,

pub fn new(store: Store, secret: &[u8]) -> SessionLayer<Store>

Creates a layer which will attach a SessionHandle to requests via an extension. This session is derived from a cryptographically signed cookie. When the client sends a valid, known cookie then the session is hydrated from this. Otherwise a new cookie is created and returned in the response.

The default behaviour is to enable “guest” sessions with PersistencePolicy::Always.

Panics

SessionLayer::new will panic if the secret is less than 64 bytes.

Customization

The configuration of the session may be adjusted according to the needs of your application:

SessionLayer::new(
    MemoryStore::new(),
    b"please do not hardcode your secret; instead use a
    cryptographically secure value",
)
.with_cookie_name("your.cookie.name")
.with_cookie_path("/some/path")
.with_cookie_domain("www.example.com")
.with_same_site_policy(SameSite::Lax)
.with_session_ttl(Some(Duration::from_secs(60 * 5)))
.with_persistence_policy(PersistencePolicy::Always)
.with_http_only(true)
.with_secure(true);

pub fn with_persistence_policy( self, policy: PersistencePolicy ) -> SessionLayer<Store>

When true, a session cookie will always be set. When false the session data must be modified in order for it to be set. Defaults to true.

pub fn with_cookie_path( self, cookie_path: impl AsRef<str> ) -> SessionLayer<Store>

Sets a cookie for the session. Defaults to "/".

pub fn with_cookie_name( self, cookie_name: impl AsRef<str> ) -> SessionLayer<Store>

Sets a cookie name for the session. Defaults to "sid".

pub fn with_cookie_domain( self, cookie_domain: impl AsRef<str> ) -> SessionLayer<Store>

Sets a cookie domain for the session. Defaults to None.

pub fn with_same_site_policy(self, policy: SameSite) -> SessionLayer<Store>

Sets a cookie same site policy for the session. Defaults to SameSite::Strict.

pub fn with_session_ttl( self, session_ttl: Option<Duration> ) -> SessionLayer<Store>

Sets a cookie time-to-live (ttl) for the session. Defaults to Duration::from_secs(60 * 60 * 24); one day.

pub fn with_http_only(self, http_only: bool) -> SessionLayer<Store>

Sets a cookie HttpOnly attribute for the session. Defaults to true.

pub fn with_secure(self, secure: bool) -> SessionLayer<Store>

Sets a cookie secure attribute for the session. Defaults to true.