Crate mailpot_web

Re-exports

pub use cal::calendarize;

pub use typed_paths::tsr::RouterExt;

pub use auth::*;

pub use cal::*;

pub use help::*;

pub use lists::*;

pub use minijinja_utils::*;

pub use settings::*;

pub use topics::*;

pub use typed_paths::*;

pub use utils::*;

Modules

auth

auth_impls 🔒

cal

help

lists

minijinja_utils

Utils for templates with the minijinja crate.

settings

topics

typed_paths

utils

Structs

AppState

AuthLayer

Layer that provides session-based authentication via AuthContext.

AuthMemoryStore

An ephemeral store, useful for testing and demonstration purposes.

Extension

Extractor and response for extensions.

Form

URL encoded extractor and response.

Html

An HTML response.

MemoryStore

in-memory session store

Path

Extractor that will get captures from the URL and parse them using serde.

Query

Extractor that deserializes query strings into some type.

ReadableSession

An extractor which provides a readable session. Sessions may have many readers.

Redirect

Response that redirects the request to another location.

Request

Represents an HTTP request.

RequireAuthorizationLayer

A wrapper around tower_http::auth::RequireAuthorizationLayer which provides login authorization.

Response

Represents an HTTP response

ResponseError

Router

The router type for composing handlers and services.

SessionLayer

Layer that provides cookie-based sessions.

State

Extractor for state.

StatusCode

An HTTP status code (status-code in RFC 7230 et al.).

WritableSession

An extractor which provides a writable session. Sessions may have only one writer.

Enums

Result

Result is a type that represents either success (Ok) or failure (Err).

Constants

BUILD_INFO

CLI_INFO

VERSION_INFO

_PACKAGE_COMMIT_SHA 🔒

Traits

AuthUser

A trait which defines methods that allow an arbitrary user type to be authenticated.

Handler

Trait for async functions that can be used to handle requests.

IntoResponse

Trait for generating responses.

IntoResponseError

IntoResponseErrorResult

TypedPath

A type safe path.

Functions

_get_package_git_sha 🔒

get

Route GET requests to the given handler.

get_git_sha

post

Route POST requests to the given handler.

Type Definitions

AuthContext

RequireAuth

SecretVec

Vec types containing secret value

Derive Macros

TypedPath

Derive an implementation of axum_extra::routing::TypedPath.